+++ 002-wrangle-RAPIDS-git-repos/nvidia_rapids_repo_wrangler.py
@@ -4,13 +4,15 @@
 # The purpose of this application to do do some discovery
 # on the composition of the RAPIDS repositories.
 #
+# One main purpose of discovery is to identify possible CI/CD integrations across all repos.
+#
 ###
 import os
 import json
 
-# given an extension map, output the top max_top_records based a zero padding specified
-def output_simple_metrics_from_extension_map(extension_map, zero_padding, max_top_records, title):
-  # output statistics on ALL the files across all repos
+# given a map, output the top max_top_records based a zero padding specified, cru
+def output_simple_metrics_from_map(extension_map, zero_padding, max_top_records, title):
+  # output statistics on the data contained in extensions_map
   num_and_ext_count_str = []
   [num_and_ext_count_str.append(f"{len(extension_map[key]):0{zero_padding}}|{key}") for key in extension_map.keys()]
   sorted_num_and_ext_count_str = sorted(num_and_ext_count_str, reverse=True)[0:max_top_records]
@@ -43,27 +45,144 @@ def handle_repo_info(extension_map, file_name):
     if file_extension not in extension_map:
         extension_map[file_extension] = []
 
+    # if we were interested in JUST counts using a unique set with incrementing counts
+    # would make sense. But it's relatively small overhead to keep a full list of all files.
+    # Might come in handy later.
     extension_map[file_extension].append(file_name)
 
-def wrangle_repo(repos_dir, repo_name, all_repo_info_by_file_type, count_top_n_extensions):
+# jenkins  ~ typical integration contains "Jenkinsfile" or naming variant
+#
+# https://www.jenkins.io/doc/book/pipeline/jenkinsfile/
+#
+# examples found in the non_file_ext_list.txt file:
+#   Jenkinsfile
+#   Jenkinsfile-win64
+def possible_jenkins_integration(file_name):
+  jenkinsfile_name = None
+
+  if 'Jenkinsfile' in file_name:
+    jenkinsfile_name = file_name
+
+  return jenkinsfile_name
+
+# circleci ~ typical integration includes `.circleci/config.yml` folder + file combination
+#
+# https://circleci.com/docs/2.0/configuration-reference/
+#
+def possible_circleci_integration(full_file_path, file_name):
+  circleci_file_name = None
+
+  if '.circleci/' in full_file_path:
+    circleci_file_name = file_name
+
+  return circleci_file_name
+
+# travisci ~ typical integration includes a `.travis.yml` file
+#
+# https://docs.travis-ci.com/user/tutorial/
+#
+# Ran into this while interviewing about the CI/CD build for >>>-- apache arrow --> ;)
+#
+# https://github.com/apache/arrow/blob/master/.travis.yml
+#
+def possible_travisci_integration(file_name):
+  travisci_file_name = None
+
+  if '.travis.yml' in file_name:
+    travisci_file_name = file_name
+
+  return travisci_file_name
+
+# github actions ~ typical integration includes a `.github/workflows` directory structure
+#
+# https://docs.github.com/en/actions/learn-github-actions/understanding-github-actions#create-an-example-workflow
+#
+def possible_github_actions_integration(full_file_path, file_name):
+  github_actions_file_name = None
+
+  if '.github/workflows' in full_file_path:
+    github_actions_file_name = file_name
+
+  return github_actions_file_name
+
+# let's add in an integration file by ci/cd type and repo
+def process_ci_cd_integration(integration_key, repo_name, file_name, ci_cd_integrations):
+      if integration_key not in ci_cd_integrations:
+        ci_cd_integrations[integration_key] = {}
+
+      if repo_name not in ci_cd_integrations[integration_key]:
+                ci_cd_integrations[integration_key][repo_name] = []
+
+      ci_cd_integrations[integration_key][repo_name].append(file_name)
+
+# let's use some common patterns to identify possible CI/CD integrations
+def identify_ci_cd_integrations(repo_name, full_file_path, file_name, ci_cd_integrations):
+    # if we detect an integration, we will return the filename, otherwise return None so we can skip it
+    if possible_jenkins_integration(file_name):
+      process_ci_cd_integration('ci-cd-jenkins', repo_name, file_name, ci_cd_integrations)
+
+    if possible_circleci_integration(full_file_path, file_name):
+      process_ci_cd_integration('ci-cd-circleci', repo_name, file_name, ci_cd_integrations)
+
+    if possible_travisci_integration(file_name):
+      process_ci_cd_integration('ci-cd-travisci', repo_name, file_name, ci_cd_integrations)
+
+    if possible_github_actions_integration(full_file_path, file_name):
+      process_ci_cd_integration('ci-cd-github-actions', repo_name, file_name, ci_cd_integrations)
+
+def output_ci_cd_integrations(ci_cd_integrations):
+  print(f"{json.dumps(ci_cd_integrations, sort_keys=True, indent=2)}")
+
+  # now for a quick summary of integrations + number of repos
+  output_simple_metrics_from_map(ci_cd_integrations, 4, len(ci_cd_integrations.keys()), 'CI/CD Integrations')
+
+def wrangle_repo(repos_dir, repo_name, all_repo_info_by_file_type, count_top_n_extensions, all_repo_ci_cd_integrations):
   # let's use some os.walk to rip through the files
   # https://stackoverflow.com/questions/19587118/iterating-through-directories-with-python
   # recursion baby!
   full_repo_path = f"{repos_dir}/{repo_name}"
   repo_info_by_file_type = {}
+  repo_ci_cd_integrations = {}
+
+  # TODO: remove to process all repos
+  # if repo_name != 'cudf':
+  #   return
 
   for subdir, dirs, files in os.walk(full_repo_path):
-    for file in files:
-        #print(os.path.join(subdir, file))
+    for file_name in files:
+        full_file_path = os.path.join(subdir, file_name)
+        #print(full_file_path)
+
+        split_subdir = subdir.split("/")
+
+        # let's skip the .git folder
+        if len(split_subdir) > 3 and split_subdir[3] == '.git':
+            continue
 
         # handle the local per-repo, extension map
-        handle_repo_info(repo_info_by_file_type, file)
+        handle_repo_info(repo_info_by_file_type, file_name)
 
         # handle the global repo extension map
-        handle_repo_info(all_repo_info_by_file_type, file)
+        handle_repo_info(all_repo_info_by_file_type, file_name)
+
+        # identify possible CI/CD interations for the repo
+        identify_ci_cd_integrations(repo_name, full_file_path, file_name, repo_ci_cd_integrations)
+
+        # identify possible CI/CD integrations across repos
+        identify_ci_cd_integrations(repo_name, full_file_path, file_name, all_repo_ci_cd_integrations)
 
   # output the summary information for the local repository
-  sorted_num_and_ext_count_str = output_simple_metrics_from_extension_map(repo_info_by_file_type, 6, 3, repo_name)
+  sorted_num_and_ext_count_str = output_simple_metrics_from_map(repo_info_by_file_type, 6, 3, repo_name)
+
+  # output the CI/CD integration information for the local repository
+  if len(repo_ci_cd_integrations) > 0:
+    output_ci_cd_integrations(repo_ci_cd_integrations)
+  # let's add in a `no_ci_cd_integrations` key and add this repo into it
+  else:
+    no_ci_cd_integrations_key = 'no_ci_cd_integrations'
+
+    process_ci_cd_integration('no_ci_cd_integrations', repo_name, '', repo_ci_cd_integrations)
+    process_ci_cd_integration('no_ci_cd_integrations', repo_name, '', all_repo_ci_cd_integrations)
 
   # create a list for the top N we have selected above, then count frequency by extension
   for i in list(range(0,len(sorted_num_and_ext_count_str))):
@@ -95,26 +214,38 @@ def show_count_of_repos_at_top_n(count_top_n_extensions):
 
     [print(f"{rank_item}") for rank_item in sorted(ranked_extension_frequency, reverse=True)[0:3]]
 
-
 # main application entrypoint
 if __name__ == "__main__":
 
   # TODO: parametize these
   repos_dir = '../repos'
 
-  # gather some metrics accross ALL repos for final analysis
-  all_repo_info_by_file_type = {}
+  # gather some metrics across ALL repos for final analysis
   count_top_n_extensions = {}
+  all_repo_info_by_file_type = {}
+  all_repo_ci_cd_integrations = {}
 
   cloned_repo_dir_names = sorted(os.listdir(repos_dir))
 
   # rip through each repo we downloaded and apply some logic
-  [wrangle_repo(repos_dir, repo_name, all_repo_info_by_file_type, count_top_n_extensions) for repo_name in cloned_repo_dir_names]
+  [wrangle_repo(repos_dir, repo_name, all_repo_info_by_file_type, count_top_n_extensions, all_repo_ci_cd_integrations) for repo_name in cloned_repo_dir_names]
 
   # output the summary information for all repositories combined
-  output_simple_metrics_from_extension_map(all_repo_info_by_file_type, 6, 10, 'TOTALS')
+  output_simple_metrics_from_map(all_repo_info_by_file_type, 6, 10, 'TOTALS')
 
   print("Count of repos with the top extension for top N")
   print(f"{json.dumps(count_top_n_extensions, sort_keys=True, indent=2)}")
 
   show_count_of_repos_at_top_n(count_top_n_extensions)
+
+  # let's dump the unique list of no_file_ext across all repos
+  no_ext_file_name = 'no_file_ext_list.txt'
+  sorted_unique_no_ext_filenames = sorted(set(all_repo_info_by_file_type['no_file_ext']))
+
+  print(f"Writing all {len(sorted_unique_no_ext_filenames)} unique non-extension files to {no_ext_file_name}")
+  with open(no_ext_file_name, 'w') as out_file:
+    [out_file.write(f"{extension}\n") for extension in sorted_unique_no_ext_filenames]
+
+  # summary of CI/CD integrations by repo
+  if len(all_repo_ci_cd_integrations) > 0:
+    output_ci_cd_integrations(all_repo_ci_cd_integrations)